Closes https://github.com/valdisiljuconoks/LocalizationProvider/issues/203
