+++ src/DbLocalizationProvider.Abstractions/DiscoveredResource.cs
