+++ src/DbLocalizationProvider.Abstractions/DiscoveredTranslation.cs
