+++ src/DbLocalizationProvider.Storage.SqlServer/ConfigurationContextExtensions.cs
@@ -29,11 +29,8 @@ namespace DbLocalizationProvider.Storage.SqlServer
             }
 
             Settings.DbContextConnectionString = connectionString;
-
             context.TypeFactory.AddTransient<IResourceRepository, ResourceRepository>();
-
             context.TypeFactory.ForQuery<UpdateSchema.Command>().SetHandler<SchemaUpdater>();
-            context.TypeFactory.ForQuery<SyncResources.Query>().SetHandler<ResourceSynchronizer>();
 
             return context;
         }
